+minor fixes
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@1272 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/auth.c
@@ -157,7 +157,7 @@ int generate_salt(char *salt, size_t max_size)
 	
 	enum e_set { set_al = 0, set_alu, set_num, set_additional, set_max };
 	
-	int i; for (i = empty_len; i < max_size - 1; ++i)
+	size_t i; for (i = empty_len; i < max_size - 1; ++i)
 	{
 		char ch = '\0';
